+++ lib/recls/obsolete.rb
@@ -44,16 +44,19 @@ if not defined? RECLS_NO_OBSOLETE
 
 	module Recls # :nodoc: all
 
+		# @!visibility private
 		def self.pathNameSeparator
 
 			PATH_NAME_SEPARATOR
 		end
 
+		# @!visibility private
 		def self.pathSeparator
 
 			PATH_SEPARATOR
 		end
 
+		# @!visibility private
 		def self.wildcardsAll
 
 			WILDCARDS_ALL
@@ -61,27 +64,43 @@ if not defined? RECLS_NO_OBSOLETE
 
 		class FileSearch # :nodoc:
 
+			# @!visibility private
 			alias_method :searchRoot, :search_root
+			# @!visibility private
 			alias_method :pattern, :patterns
 		end
 
 		class Entry # :nodoc:
 
+			# @!visibility private
 			alias_method :uncDrive, :drive
+			# @!visibility private
 			alias_method :directoryPath, :directory_path
+			# @!visibility private
 			alias_method :directoryParts, :directory_parts
+			# @!visibility private
 			alias_method :file, :file_full_name
+			# @!visibility private
 			alias_method :shortFile, :file_short_name
+			# @!visibility private
 			alias_method :fileBaseName, :file_name_only
+			# @!visibility private
 			alias_method :fileName, :file_name_only
+			# @!visibility private
 			alias_method :fileExt, :file_extension
+			# @!visibility private
 			alias_method :searchDirectory, :search_directory
+			# @!visibility private
 			alias_method :searchRelativePath, :search_relative_path
 
+			# @!visibility private
 			alias_method :isDirectory, :directory?
+			# @!visibility private
 			alias_method :isFile, :file?
 			#alias_method :isLink, :link?
+			# @!visibility private
 			alias_method :isReadOnly, :readonly?
+			# @!visibility private
 			def isUNC
 
 				d = drive
@@ -89,6 +108,7 @@ if not defined? RECLS_NO_OBSOLETE
 				d and d.size > 2
 			end
 
+			# @!visibility private
 			alias_method :creationTime, :modification_time
 		end
 	end # module Recls