+++ CHANGES.md
@@ -1,5 +1,11 @@
 # **recls.Ruby** Changes
 
+## 2.10.1 - 14th April 2019
+
+* + added examples/find_files_and_directories.rb
+* ~ making examples executable
+* ~ hiding private/impl constructs from YARD
+
 ## 2.10.0 - 14th April 2019
 
 * + added Recls::absolute_path?